Updates for gradle 9.4.1 and kotlin 2.4. Change com.github.alpbak:BoxedVerticalSeekBar:1.1.1 for com.github.ML-SPD:BoxedVerticalSeekBar:v1.1.1-fix5 due to some compilation errors. Remove some useless code.

+++ app/src/main/java/org/gnu/itsmoroto/midandpad/EventButton.kt
@@ -393,9 +393,9 @@ class EventButton : androidx.appcompat.widget.AppCompatButton {
         when (mType) {
             MidiHelper.EventTypes.EVENT_NOTE -> {
                 command = if (type == ON)
-                    MidiHelper.STATUS_NOTE_ON or channel.toUByte()
+                    MidiHelper.STATUS_NOTE_ON or channel
                 else
-                    MidiHelper.STATUS_NOTE_OFF or channel.toUByte()
+                    MidiHelper.STATUS_NOTE_OFF or channel
 
                 msg.add(mNoteNumber.toUByte())
                 msg.add(vel.toUByte())
@@ -403,9 +403,9 @@ class EventButton : androidx.appcompat.widget.AppCompatButton {
 
             MidiHelper.EventTypes.EVENT_CHORD -> {
                 command = if (type == ON)
-                    MidiHelper.STATUS_NOTE_ON or channel.toUByte()
+                    MidiHelper.STATUS_NOTE_ON or channel
                 else
-                    MidiHelper.STATUS_NOTE_OFF or channel.toUByte()
+                    MidiHelper.STATUS_NOTE_OFF or channel
 
                 for (note in mChordNotes) {
                     msg.add(note.toUByte())
